+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Changed
 - Build an iframe and change the domain in order to avoid preflight checks.
-- Moved the WebSocket exports into their own named exports and out of the main client.
+- Moved the WebSocket exports into their own named exports and out of the main client: `onEventStarting`, `onLeaderboardUpdated`, `onNotificationReceived`, `onRecognitionCreated`, and `onUserActivityCreated`.
 - Moved the `autoPaginate` function into its own named export.
 - Moved `endUserSimulation`, `isSimulating`, and `startUserSimulation` functions into their own named exports.
 - Moved the signed in state functions into their own named exports: `isSignedIn`, `setToken`, `signIn`, and `signOut`.